+++ sc-kpm/sc-agents-common/utils/IteratorUtils.cpp
@@ -69,28 +69,20 @@ ScAddr IteratorUtils::getNextFromSet(
   SC_CHECK_PARAM(previous, ("Invalid previous element address"));
   SC_CHECK_PARAM(sequenceRelation, ("Invalid sequence relation address"));
 
-  std::string const NEXT_ELEMENT_ALIAS = "_next_element";
-  std::string const NEXT_ELEMENT_ACCESS_ARC_ALIAS = "_next_element_access_arc";
-  std::string const PREVIOUS_ELEMENT_ACCESS_ARC_ALIAS = "_previous_element_access_arc";
-
-  ScTemplate scTemplate;
-  scTemplate.Triple(set, ScType::EdgeAccessVarPosPerm >> PREVIOUS_ELEMENT_ACCESS_ARC_ALIAS, previous);
-  scTemplate.Triple(
-      set, ScType::EdgeAccessVarPosPerm >> NEXT_ELEMENT_ACCESS_ARC_ALIAS, ScType::Unknown >> NEXT_ELEMENT_ALIAS);
-  scTemplate.TripleWithRelation(
-      PREVIOUS_ELEMENT_ACCESS_ARC_ALIAS,
-      ScType::EdgeDCommonVar,
-      NEXT_ELEMENT_ACCESS_ARC_ALIAS,
-      ScType::EdgeAccessVarPosPerm,
-      sequenceRelation);
-  ScTemplateSearchResult searchResult;
-  ms_context->HelperSearchTemplate(scTemplate, searchResult);
-
-  if (searchResult.Size() == 1)
+  ScAddr nextElement;
+  ScIterator3Ptr const & previousElementIterator = ms_context->Iterator3(set, ScType::EdgeAccessConstPosPerm, previous);
+  if (previousElementIterator->Next())
   {
-    return searchResult[0][NEXT_ELEMENT_ALIAS];
+    ScAddr const & previousElementEdge = previousElementIterator->Get(1);
+    ScAddr const & nextElementEdge =
+        IteratorUtils::getAnyByOutRelation(ms_context, previousElementEdge, sequenceRelation);
+    if (nextElementEdge.IsValid())
+    {
+      nextElement = ms_context->GetEdgeTarget(nextElementEdge);
+    }
   }
-  return {};
+
+  return nextElement;
 }
 
 ScAddrVector IteratorUtils::getAllWithType(ScMemoryContext * ms_context, const ScAddr & set, ScType scType)